+++ packages/menu/src/Menu/index.ts
@@ -1,2 +1,2 @@
 export { default as Menu } from './Menu';
-export type { MenuProps } from './Menu.types';
+export { type MenuProps, MenuVariant } from './Menu.types';