+++ packages/menu/src/MenuContext/MenuContext.tsx
@@ -5,6 +5,7 @@ import { createDescendantsContext } from '@leafygreen-ui/descendants';
 import { Theme } from '@leafygreen-ui/lib';
 
 import { HighlightReducerReturnType } from '../HighlightReducer/highlight.types';
+import { MenuVariant } from '../Menu/Menu.types';
 import { getLgIds, GetLgIdsReturnType } from '../utils';
 
 export interface MenuContextData extends HighlightReducerReturnType {
@@ -18,6 +19,11 @@ export interface MenuContextData extends HighlightReducerReturnType {
    * LGIDs for menu components.
    */
   lgIds: GetLgIdsReturnType;
+
+  /**
+   * Variant of the menu to be rendered.
+   */
+  variant?: MenuVariant;
 }
 
 /**
@@ -34,6 +40,7 @@ export const MenuContext = createContext<MenuContextData>({
   moveHighlight: noop,
   setHighlight: noop,
   lgIds: getLgIds(),
+  variant: MenuVariant.Default,
 });
 
 /**