+++ chat/avatar/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @lg-chat/avatar
 
+## 6.0.0
+
+### Patch Changes
+
+- Updated dependencies [3e8fbd7]
+- Updated dependencies [82d0ad2]
+  - @leafygreen-ui/avatar@3.1.0
+  - @lg-chat/leafygreen-chat-provider@4.1.0
+
 ## 5.0.3
 
 ### Patch Changes